searchfilters: dynamically generate UIs and evaluate selected sort and content filters
SearchFilterLogic.java:
=======================
This class handles all the user interaction with the content and sort filters
of NewPipeExtractor.

The class works standalone to just get the default selected filters eg.
during init phase. See in SearchFragment#initializeFilterData()

BaseSearchFilterUiGenerator.java:
=================================
It extends SearchFilterLogic and is used as a base class to implement the UI interface
for content and sort filter dialogs eg. SearchFilterDialogGenerator or
SearchFilterOptionMenuAlikeDialogGenerator.