searchfilters: make default dialog not stretch vertically
Originally stretching vertically over the whole screen was used
to avoid resizing the UI each time the selected content filter
selects another set of sort filters.

As it turns out fullscreen vertically dialogs can be ugly and my
other attempt to use View.INVISIBLE instead of (View.GONE) in order
to have some sort of placeholder, didn't work well with the Spinner
Views either. So we go back and let Android resize the UI.

Another resizing annoyance is now avoided. The Material's Chip with the
Filter style was also resized each time a Chip was selected.

To avoid resizing for the Chip views the checkmark icon is removed
by using a custom chip style: "@style/ChipSearchFilter". The
custom chip style also improves the contrast when a chip is selected

To save some space a ChipGroup that holds more than
CHIP_GROUP_ELEMENTS_THRESHOLD elements will be spanned
over all cells of a row in the GridLayout.